+++ src/main/java/nqueens/Main.java
@@ -2,7 +2,7 @@ package nqueens;
 
 public class Main {
     public static void main(String[] args) {
-	NaiveSolver ns = new NaiveSolver(8);
-	ns.printSolutions();
+	BruteForceSolver solver = new BruteForceSolver(8,false);
+	System.out.println(solver.solutionsToString());
     }
 }