+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2018-06-10
+### Added
+* `Utils` to hold some generic, static methods: `copy`, `arrayToArrayList`, `threePointsFormALine`.
+* Add tests for `noThreeInLineConstraint`.
+### Changed
+* Rename `NaiveSolver` to `BruteForceSolver`.
+* Add parameter to `BruteForceSolver` constructor `noThreeInLineConstraint`.
+
 ## [0.1.0] - 2018-06-09
 ### Added
 * Implement a naive solution (disregarding performance or space) for the N-Queens problem in NaiveSolution.